redesign tha homepage navbar
