+++ components/common/root_nav/root_nav.css
@@ -1,12 +1,3 @@
-.header {
-  position: fixed;
-  width: 100%;
-  height: auto;
-  box-sizing: border-box;
-  z-index: 101;
-  color: white;
-}
-
 .background {
   background-color: #352786;
   opacity: 0.2;
@@ -25,9 +16,9 @@
   position: relative;
 }
 
-.bar p {
+/* .bar p {
   margin: 0px;
-}
+} */
 
 .bar a {
   text-decoration: none;
@@ -35,7 +26,7 @@
 
 .bar .el {
   border-radius: 50%;
-  background-color: #455ce9;
+  background-color: white;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -55,7 +46,7 @@
 }
 
 .bar .el .burger {
-  width: 30px;
+  width: 25px;
   position: relative;
   pointer-events: none;
 }
@@ -63,9 +54,9 @@
 .bar .el .burger::after,
 .bar .el .burger::before {
   content: "";
-  height: 1px;
+  height: 2px;
   width: 100%;
-  background-color: white;
+  background-color: black;
   position: relative;
   display: block;
   transition: all 1s cubic-bezier(0.76, 0, 0.24, 1);
